Various Changes
Changed naming convention on BP Equipment to better distinguish between other assets of the same name.
Seperated PickUpInteract Interface into BPI_PickUp and BPI_UseInteract Interfaces. This is so the Pick Up and Drop part of equipment can be individually added to objects in the world without interaction functionality and hence won't be checked for it.
Added some "Flame" assets in an attempt to get a working flame for lighter/candle, but is a work in progress.
(For some reason it "added" the BPs I renamed again, instead of just a modify)